\name{OrderedCategoricalGLM-package}
\alias{OrderedCategoricalGLM-package}
\alias{OrderedCategoricalGLM}
\docType{package}
\title{
\packageTitle{OrderedCategoricalGLM}
}
\description{
\packageDescription{OrderedCategoricalGLM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OrderedCategoricalGLM}
\packageIndices{OrderedCategoricalGLM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OrderedCategoricalGLM}

Maintainer: \packageMaintainer{OrderedCategoricalGLM}
}
